Temporary mitigation for SQLITE_BUSY errors.
It seems that SQLite will throw errors at any point within a transaction upon write contention. I am not planning on migrating to another embedded DB and certainly not to a standalone engine for a program this small, so an ugly mitigation will have to do for now.

Solution will be to make DB access patterns single threaded.